feat: F4 + F5 cleanup — wire document_symbols, tighten dead_code annotations (#86)
* feat: F4 + F5 cleanup — wire symbols, tighten dead_code annotations

F4:
- Wire features/symbols.rs to document_symbol_impl (backend is now a
  thin 8-line adapter; mapping logic lives in compute_document_symbols)
- Remove #[allow(dead_code)] from pub(crate) mod symbols

F5 cleanup:
- Remove blanket #[allow(dead_code)] from mod traits_impl (unused lint
  does not fire — impls are used via generic bounds)
- Remove blanket #[allow(dead_code)] from all trait blocks in traits.rs;
  annotate only the 4 genuinely-unused API methods (for_each_indexed_file,
  lines_for, word_at, word_and_range_at) which are planned for future
  workspace_symbols / CLI extraction


* feat: extract workspace_symbols to features layer

- Create features/workspace_symbols.rs with compute_workspace_symbols()
  taking SymbolIndex + SearchAccess bounds
- WorkspaceSymbolQuery, collect_index_symbols, rg_symbol_search,
  symbol_information helpers all move out of handlers.rs
- symbol_impl becomes a 5-line adapter
- traits_impl: for_each_indexed_file delegates to the concrete method
  (was duplicating the loop body)
- Removes #[allow(dead_code)] on for_each_indexed_file trait method

handlers.rs is now 128 lines (was 258 before this session)


* feat: extract completion_resolve to features/completion.rs

- Add resolve_completion_item<I: IndexRead>() to features/completion.rs
- completion_resolve in mod.rs becomes a 3-line adapter
- mod.rs: 731 → 696 lines


* refactor: replace hardcoded completion data keys with named constants

Introduce DATA_URI/DATA_LINE/DATA_COL/DATA_CALLING_URI in
features/completion.rs as the single source of truth for the
CompletionItem.data JSON blob keys used between the write side
(resolver/complete.rs) and the read side (features/completion.rs).


* fix: address PR #86 review comments

- Eliminate double file_symbols() clone: document_symbol_impl now fetches
  symbols once, skips the on-demand re-index when non-empty, re-fetches
  only after indexing. compute_document_symbols takes Vec<SymbolEntry>.
- Fix DATA_* constant layer direction: constants move to resolver/complete.rs
  (write side, lower layer) and are re-exported from features/completion.rs
  (read side). Removes the resolver->features inversion.
- features/traits.rs: annotate file_symbols #[allow(dead_code)] since
  it's now accessed via concrete Indexer, not through the trait.


---------
