+++ src/resolver/complete.rs
@@ -14,6 +14,17 @@ use super::{
     already_imported, ensure_file_data, fqns_for_name, resolve_symbol_no_rg, walk_hierarchy,
 };
 
+// ─── CompletionItem.data JSON keys ───────────────────────────────────────────
+
+/// Symbol definition URI.
+pub(crate) const DATA_URI: &str = "u";
+/// Symbol definition line (0-based).
+pub(crate) const DATA_LINE: &str = "l";
+/// Symbol definition UTF-16 column (0-based).
+pub(crate) const DATA_COL: &str = "c";
+/// Calling-site URI, present only for cross-file substitution context.
+pub(crate) const DATA_CALLING_URI: &str = "cu";
+
 // ─── match scoring ────────────────────────────────────────────────────────────
 
 /// Returns true if `name` is SCREAMING_SNAKE_CASE (all letters are uppercase).
@@ -585,9 +596,9 @@ fn completion_item_for_nested_symbol(
         ),
         None => signature,
     });
-    let mut data = serde_json::json!({"u": uri_str, "l": s.selection_start(), "c": s.selection_range.start.character});
+    let mut data = serde_json::json!({DATA_URI: uri_str, DATA_LINE: s.selection_start(), DATA_COL: s.selection_range.start.character});
     if let Some(calling_uri) = caller.uri {
-        data["cu"] = serde_json::Value::String(calling_uri.to_owned());
+        data[DATA_CALLING_URI] = serde_json::Value::String(calling_uri.to_owned());
     }
     CompletionItem {
         label: s.name.clone(),
@@ -887,7 +898,7 @@ impl<'a> BareCompletionWalk<'a> {
                 0,
                 self.prefix,
                 &symbol.detail,
-                Some(serde_json::json!({"u": self.from_uri.as_str(), "l": symbol.selection_start(), "c": symbol.selection_range.start.character})),
+                Some(serde_json::json!({DATA_URI: self.from_uri.as_str(), DATA_LINE: symbol.selection_start(), DATA_COL: symbol.selection_range.start.character})),
             );
         }
 
@@ -927,7 +938,7 @@ impl<'a> BareCompletionWalk<'a> {
                     1,
                     self.prefix,
                     &symbol.detail,
-                    Some(serde_json::json!({"u": package_uri.as_str(), "l": symbol.selection_start(), "c": symbol.selection_range.start.character})),
+                    Some(serde_json::json!({DATA_URI: package_uri.as_str(), DATA_LINE: symbol.selection_start(), DATA_COL: symbol.selection_range.start.character})),
                 );
             }
         }
